requirements.txt
pycryptodome
pytest
pytest-cov